pyproject.toml
[tool.poetry]
name = "tournament-user-service"
version = "0.1.0"
description = ""
authors = ["Connor Richmond-Clark <crichmondclark@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.108.0"
piccolo = {extras = ["all"], version = "^1.2.0"}
pydantic = "^2.5.3"
ruff = "^0.1.9"
python-jwt = "^4.1.0"
isort = "^5.13.2"
sqlmodel = "^0.0.14"
alembic = "^1.13.1"
sqlalchemy-libsql = "^0.1.0"
python-dotenv = "^1.0.0"
grpcio-tools = "^1.60.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
